Prefer double_quotes("") over single_quotes('')
String literals can be either and I don't really care as long as it's
consistent.

So why the change?
When working with other programming languages, most of them enforce
double quotes. This therefore allows for more consistency across
languages as well as muscle memory when typing in any languages.

After getting used to single quotes for a bit in ruby, when switching to
another language, I was often having to hit <backpace> and change it
back to a double quote as the other language would not support single
quote.